fix stuff

+++ src/lib/db/utils.ts
@@ -1,8 +1,11 @@
-import { sql, transaction } from '~/lib/db'
+import { isServer } from 'solid-js/web'
 import { fileSchema, folderSchema } from '~/types'
+import { getSQLocalDB } from '.'
 
 export const query = {
 	async getFiles(parent_id?: number) {
+		if (isServer) return []
+		const { sql } = await getSQLocalDB()
 		return fileSchema
 			.array()
 			.parse(
@@ -12,6 +15,8 @@ export const query = {
 			)
 	},
 	async getFolders(parent_id?: number) {
+		if (isServer) return []
+		const { sql } = await getSQLocalDB()
 		return folderSchema
 			.array()
 			.parse(
@@ -21,11 +26,15 @@ export const query = {
 			)
 	},
 	async getFile(id: number) {
+		if (isServer) return {} as TFile
+		const { sql } = await getSQLocalDB()
 		return fileSchema.parse(
 			await sql`SELECT * FROM files WHERE id = ${id}`.then((result) => result[0])
 		)
 	},
 	async getFolder(id: number) {
+		if (isServer) return {} as TFolder
+		const { sql } = await getSQLocalDB()
 		return folderSchema.parse(
 			await sql`SELECT * FROM folders WHERE id = ${id}`.then((result) => result[0])
 		)
@@ -38,6 +47,8 @@ export const query = {
 		return { files, folders }
 	},
 	async getFileByNameAndParentId(name: string, parent_id: number) {
+		if (isServer) return {} as TFile
+		const { sql } = await getSQLocalDB()
 		return fileSchema.parse(
 			await sql`SELECT * FROM files WHERE name = ${name} AND parent_id = ${parent_id}`.then(
 				(result) => result[0]
@@ -45,6 +56,8 @@ export const query = {
 		)
 	},
 	async getFolderByNameAndParentId(name: string, parent_id: number) {
+		if (isServer) return {} as TFolder
+		const { sql } = await getSQLocalDB()
 		return folderSchema.parse(
 			await sql`SELECT * FROM folders WHERE name = ${name} AND parent_id = ${parent_id}`.then(
 				(result) => result[0]
@@ -55,20 +68,28 @@ export const query = {
 
 export const mutations = {
 	async addContact(name: string, phone: string) {
-		return sql`INSERT INTO contacts (name, phone) VALUES (${name}, ${phone})`
+		if (isServer) return
+		const { sql } = await getSQLocalDB()
+		await sql`INSERT INTO contacts (name, phone) VALUES (${name}, ${phone})`
 	},
 	async addFolder(name: string, parent_id?: number) {
-		return parent_id
+		if (isServer) return
+		const { sql } = await getSQLocalDB()
+		parent_id
 			? await sql`INSERT INTO folders (name, parent_id) VALUES (${name}, ${parent_id})`
 			: await sql`INSERT INTO folders (name, parent_id) VALUES (${name}, (SELECT id FROM folders WHERE name = 'root'))`
 	},
 	async addFile(name: string, parent_id: number, metadata: Record<string, unknown> = {}) {
-		return parent_id
+		if (isServer) return
+		const { sql } = await getSQLocalDB()
+		parent_id
 			? await sql`INSERT INTO files (name, parent_id, metadata) VALUES (${name}, ${parent_id}, json(${JSON.stringify(metadata)}))`
 			: await sql`INSERT INTO files (name, parent_id, metadata) VALUES (${name}, (SELECT id FROM folders WHERE name = 'root'), json(${JSON.stringify(metadata)}))`
 	},
-	deleteFolders(ids: number[]) {
-		return transaction((sql) =>
+	async deleteFolders(ids: number[]) {
+		if (isServer) return
+		const { transaction } = await getSQLocalDB()
+		await transaction((sql) =>
 			ids.flatMap((id) => [
 				sql`DELETE FROM files WHERE parent_id = ${id}`,
 				sql`DELETE FROM folders WHERE parent_id = ${id}`,
@@ -76,13 +97,19 @@ export const mutations = {
 			])
 		)
 	},
-	deleteFiles(ids: number[]) {
-		return transaction((sql) => ids.map((id) => sql`DELETE FROM files WHERE id = ${id}`))
+	async deleteFiles(ids: number[]) {
+		if (isServer) return
+		const { transaction } = await getSQLocalDB()
+		await transaction((sql) => ids.map((id) => sql`DELETE FROM files WHERE id = ${id}`))
 	},
 	async renameFolder(id: number, name: string) {
-		return sql`UPDATE folders SET name = ${name} WHERE id = ${id}`
+		if (isServer) return
+		const { sql } = await getSQLocalDB()
+		sql`UPDATE folders SET name = ${name} WHERE id = ${id}`
 	},
 	async renameFile(id: number, name: string) {
+		if (isServer) return
+		const { sql } = await getSQLocalDB()
 		return sql`UPDATE files SET name = ${name} WHERE id = ${id}`
 	}
 }
